PATCH by Rik Snel <rsnel@cube.dyndns.org>
this patch enhances the jpeg header writer. It can be asked to omit
quantisation and huffman tables and it can write different horizontal and
vertical sampling factors. (the last thing is useless for libavcodec
itself (because libavcodec only handles YUV420P at ecoder level), but the
values are initialized so that operation of libavcodec is not impaired)


git-svn-id: svn://svn.ffmpeg.org/ffmpeg/trunk/libavcodec@290 9553f0bf-9b14-0410-a0b8-cfaf0461ba5b