Rename template included sources from .h to _template.c.
There are multiple source files that are #include'd rather than
compiled, as they are used as template for generation of similar code,
like asm-optimised code. Some of these files are right now named with
a .h extension, although they are not header in any reasonable sense.

Rename them so that instead of being named with .h extension they are
named with _template.c as final part.



git-svn-id: svn://svn.ffmpeg.org/ffmpeg/trunk/libavcodec@15730 9553f0bf-9b14-0410-a0b8-cfaf0461ba5b